refactor existing backend tests
